+++ CHANGELOG.md
@@ -9,6 +9,18 @@
 ### Bug fixes
 
 * Fixed uninitialized server name in TLS config ([#93](https://github.com/microsoft/go-mssqldb/issues/93))([#94](https://github.com/microsoft/go-mssqldb/pull/94))
+* Fixed several kerberos authentication usages on Linux with new krb5 authentication provider. 
+
+### Changed
+
+* New kerberos authenticator implementation uses more explicit connection string parameters.
+
+| Old          | New                |
+|--------------|--------------------|
+| krb5conffile | krb5-configfile    |
+| krbcache     | krb5-credcachefile |
+| keytabfile   | krb5-keytabfile    |
+| realm        | krb5-realm         |
 
 ## 0.20.0